extract inner components to separate files to simplify parent component
